feat: improve tests

+++ currency-api/src/modules/jest/setupTestDB.ts
@@ -1,13 +1,15 @@
 import mongoose from 'mongoose';
 import config from '../../config/config';
 
-const setupTestDB = () => {
+const setupTestDB = (clearDB: boolean = true) => {
   beforeAll(async () => {
     await mongoose.connect(config.mongoose.url);
   });
 
   beforeEach(async () => {
-    await Promise.all(Object.values(mongoose.connection.collections).map(async (collection) => collection.deleteMany({})));
+    if(clearDB) {
+      await Promise.all(Object.values(mongoose.connection.collections).map(async (collection) => collection.deleteMany({})));
+    }
   });
 
   afterAll(async () => {
